Error fixes validated.

+++ WindowsFormsApp/FrmMain.cs
@@ -206,17 +206,21 @@ namespace WindowsFormsApp
         /// <param name="e"></param>
         private void PbxAddCustomer_click(object sender, EventArgs e)
         {
-            // Listboxes components content verification.
-            ListBoxesVerifications();
+            if (lbxCustomers.SelectedItem == null)
+            {
+                MessageBox.Show("No customer selected.", "Warning", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+            }
+            else
+            {
+                // Gets the selected custumer from lbxCustomersSelected component.
+                Customer customer = (Customer)lbxCustomers.SelectedItem;
 
-            // Gets the selected custumer from lbxCustomersSelected component.
-            Customer customer = (Customer)lbxCustomers.SelectedItem;
+                // Refreshes two ListBoxes components.
+                Refresh_listboxes(lbxCustomers, customers, customer, lbxCustomersSelected, customersSelected);
 
-            // Refreshes two ListBoxes components.
-            Refresh_listboxes(lbxCustomers, customers, customer, lbxCustomersSelected, customersSelected);
-
-            // Listboxes components content verification.
-            ListBoxesVerifications();
+                // Listboxes components content verification.
+                ListBoxesVerifications();
+            }
         }
 
         #endregion
@@ -251,8 +255,10 @@ namespace WindowsFormsApp
         /// <param name="e"></param>
         private void PbxRemoveCustomer_click(object sender, EventArgs e)
         {
-            // Listboxes components content verification.
-            ListBoxesVerifications();
+            if (lbxCustomersSelected.SelectedItem == null)
+            {
+                MessageBox.Show("No customer selected.", "Warning", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+            }
 
             // Gets the selected custumer from lbxCustomersSelected component.
             Customer customer = (Customer)lbxCustomersSelected.SelectedItem;
@@ -330,21 +336,29 @@ namespace WindowsFormsApp
         /// <param name="e"></param>
         private void PbxRemove_click(object sender, EventArgs e)
         {
-            DialogResult result = MessageBox.Show("Are you sure you want to delete this customer?", "Confirm the deletation", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
-
-            if (result == DialogResult.Yes)
+            // Controls if the database contains customer(s).
+            if (lbxCustomers.Items.Count < 1 || lbxCustomers.SelectedItem == null)
             {
-                // Gets the currently customer selected in listbox component.
-                Customer customer = (Customer)lbxCustomers.SelectedItem;
+                MessageBox.Show("No customer in database or selected.", "Warning", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+            }
+            else
+            {
+                DialogResult result = MessageBox.Show("Are you sure you want to delete this customer?", "Confirm the deletation", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
 
-                // Delete the customer in database and in the software memory.
-                access.DeleteCustomer(customer.GetId());
+                if (result == DialogResult.Yes)
+                {
+                    // Gets the currently customer selected in listbox component.
+                    Customer customer = (Customer)lbxCustomers.SelectedItem;
 
-                // Delete the customer in customers list.
-                customers.Remove(customer);
+                    // Delete the customer in database and in the software memory.
+                    access.DeleteCustomer(customer.GetId());
 
-                // Refresh the ListBox component.
-                Refresh_listBox(lbxCustomers, customers);
+                    // Delete the customer in customers list.
+                    customers.Remove(customer);
+
+                    // Refresh the ListBox component.
+                    Refresh_listBox(lbxCustomers, customers);
+                }
             }
         }
